Deleted Countdown Controller
Added it to player controller, player and ai both have to wait for the game to start before any inputs, 60 timer has been added, if time expires the player loses. Ai script updated along with player script to respond to both timers. also added a lose condition if the player loses to the AI